+++ src/Core/Kernel.php
@@ -39,5 +39,9 @@ class Kernel {
         if (!empty($mods['customer_order_note']) && class_exists('HB\\UCS\\Modules\\CustomerOrderNote\\CustomerOrderNoteModule')) {
             (new \HB\UCS\Modules\CustomerOrderNote\CustomerOrderNoteModule())->init();
         }
+
+        if (!empty($mods['checkout']) && class_exists('HB\\UCS\\Modules\\Checkout\\CheckoutModule')) {
+            (new \HB\UCS\Modules\Checkout\CheckoutModule())->init();
+        }
     }
 }